+++ test/test_helper.exs
@@ -33,6 +33,7 @@ ExUnit.start(
     # :pix_infraction_log,
     # :pix_domain,
     # :static_brcode,
+    # :dynamic_brcode,
     # :issuing_balance,
     # :issuing_product,
     # :issuing_card,
@@ -72,6 +73,7 @@ Code.require_file("./test/utils/pix_dispute.exs")
 Code.require_file("./test/utils/pix_pull_request.exs")
 Code.require_file("./test/utils/random.exs")
 Code.require_file("./test/utils/static_brcode.exs")
+Code.require_file("./test/utils/dynamic_brcode.exs")
 Code.require_file("./test/utils/ledger.exs")
 Code.require_file("./test/utils/ledger_transaction.exs")
 Code.require_file("./test/utils/issuing_holder.exs")